backend/requirements.txt
asgiref==3.8.1
certifi==2025.11.12
cffi>=2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.8.0rc2
Django==4.2.26
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
idna==3.11
oauthlib==3.3.1
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
simplejwt==2.0.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.4
typing==3.7.4.3
typing_extensions==4.13.2
urllib3==2.2.3
